requirements.txt
fastapi
uvicorn
requests
websockets
python-dotenv
psutil
onvif-zeep
cryptography
opencv-python-headless
